Refactoring: Finish API re-organization
Move remaining API endpoints to respective modules.
